pyproject.toml
[project]
name = "health"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.136.0",
    "ipykernel>=7.2.0",
    "joblib>=1.5.3",
    "jupyter>=1.1.1",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.44.0",
    "xgboost>=3.2.0",
]
